fix(phone/match): fix position for draggable cards
some t/ranslation issued resolved as well
